Maintain items through to elaboration
This reworks items so that instead of trying to group together the claims and definitions during desugaring, we instead maintain the items all the way through to elaboration.

This has the effect of preventing shadowing at the top level, and allowing for the forward declaration of definitions, and should make mutual recursion easier to implement in the future!