Add reserved types
This allows us to document reserved fields, ensuring that their values aren’t dependended on. This should ensure that forwards compatibility is maintained.

In the future we’ll want the type signature to be something like:

```
Reserved : (A : Type) (default : Repr A) -> Type
```

The `default` value will come up when encoding data, and is helpful for documentation purposes. We don’t have a `Repr` type yet, so we’ll heve to defer this for now.