Make aliases transparent
We now pass items through to the evaluator, allowing us to resolve aliases to the terms they point to. This mirrors the way that type aliases work in Rust. We might need to be careful about what happens when wrapper structs are generated in the Rust backend, but I’ll leave that to another day!

Closes #162.