Define type system as a PTS
Following from the developments in the Makam spec (see a61e3809) this
switches implementation of the type system to be closer to a standard
definition of a Pure Type System (PTS). While I don’t think we get any
metatheory for free with this presentation, it might provide a useful
signpost for type system researchers who might be interested in the
properties of our type system.

The configuration of the Pure Type System we use is as follows:

- Sorts
  - `Type`
  - `Kind`
- Axioms
  - `Type` : `Kind`
- Rules
  - (`Type`, `Type`, `Type`)
  - (`Type`, `Kind`, `Kind`)
  - (`Kind`, `Type`, `Kind`)
  - (`Kind`, `Kind`, `Kind`)

This ultimately gives us a 'truncated' version of Martin Löf Type
Theory. Since we are not intending to make a theorem prover, the two
sorts, `Type` and `Kind`, _should_ be enough for our purposes. `Format`
is not considered to be a sort, as was the case in previous
experiments. This conveniently prevents format descriptions from being
used in type annotations, which was the case in the past!