Return values from binary reader
Thanks to the addition of array terms and struct terms we are now able to represent all format representations in the host language. This lets us remove `lang::core::binary::Term`, and return values from the interpreted binary reader instead, making it easier to implement dependent parsing in the future.