Cleanup grammars in specification
This simplifies the grammars in the specification by removing the formatting. This should be easier to update. In the future we could use a MDBook plugin to format the grammars nicely in HTML.

I’ve also updated the grammars to match the current surface language (some things were missing or had changed).