Rename variants in the core language (#358)
Currently I use terminology from natural deduction to inform the names of the variants in the core language. This has turned out to be a bit intimidating and a source of confusion. The consistency between the names seems to also get in the way of understanding.

In the interest of making it a bit easier to understand and work on Fathom, this PR switches to more familiar names. We can tweak this further down the line, but I hope it helps!