Add file ids to byte ranges
This allows us to more easily handle fathom terms across multiple files, but does add a significant number of duplicated `FileId`s to `ByteRange`s. This could become expensive if we plan to add position information to the core language where it might impact the performance of the evaluator. We might need to revisit this later down the line if it ever becomes an issue.