Test the command line interface
The command line is getting a bit more complicated and I was running into issues where I’d accidentally alter its behaviour without realising, so I thought it would be good to add some tests.

I was looking for a `cram` implementation for Rust (which I’ve found pretty useful in OCaml), and `trycmd` seemed pretty close. It’s not widely used at the moment, but the crates it is used in are pretty popular.