Attempt to address TODO in elaboration `convert`
Only relying on the span in the core term revealed at least one case
where span information was not present. This was addressed by populating
the span on FlexibleInsertion.

The possibility of the span of the core term being empty is handled by
emitting a warning but continuting on using the surface range. Ideally
we'd be able to remove the surface range in the future.