Cache values in elaborator to avoid copies
This probably doesn’t do a *huge* amount to improve memory usage - we
would need but to look into the evaluator for that - but it’s a start
and makes me slightly less uncomfortable when constructing these common
values.

We could do this more systematically, for example with some sort of
“value allocator” for creating reference counted values, or perhaps
rethink our value representation, for example, representing them
“intrusively”, attempting to avoid dynamic allocation when values are
small.