Make it easier to distill/pretty print values
This reduces the gymnasics we’ve needed to do whenever we want to pretty
print something by cloning the name environment as opposed to taking it
by mutable reference when constructing a distillation environment.

This means we can no longer reuse the existing allocation when
temporarily pushing local names, but I think the savings we got from
this was probably minor, and the increase in clarity is worth it.